feat: Created restart button that stops timer per global event
A global event trigger function to stop the timer was created (triggerStopTimer).
The timer uses a setInterval to update periodically.
To be able to stop the timer based on different actions (e.g. game end or
restart button push), an event listener is created (when the timer is started)
that stops the timer interval (clearInterval) when the 'stopTimer' event
is triggered.
Other solutions would have required the passing of the timer ID through
several functions.