fix: Restart button cancels timeout for equality response
Because of the timeout set after two cards are clicked, before the response is performed
it can happen that the restart button is clicked during this delay. To prevent responses
that where triggered before restart button is clicked, being  performed after the
restart button was clicked (in the new game), the timeout has to be cleared. The
function defined in the timeout will not be performed.