feat: Changed footer padding and added minor break points
The footer text seemed a bit too close to the game on taller
viewports.

Also, limited the max width for short viewports. If these short
viewports are wide (e.g. HD ready 1280x720) the bottom portion
of the game was cut weirdly.
Now the game is forced slimmer, which reduced the card size and
brings the bottom end further up. Everything should be on the
screen now.