fix: Fixed animation timeout issues
Because of the timeouts that are defined for animations, it was possible
that multiple event listeners for "twoCardsClicked" existed.

If the Restart button was clicked while an animation was going,
some functions where delayed. These included the creation functions
for the event listeners. Once the timeout was up, the eventlisteners would be created.
In the meantime, the buildGame function had created these eventlisteners already.
Therefore, they existed twice (or more often if the processes was done again).

This would lead to multiple executions of the event function.
The twoCardsClicked event also leads to an increase of the moves counter.
Therefore, the moves counter would increase more than one with two cards being clicked.