Moved all served code to src directory
Because the index.html was located in the repo root, the root directory
was directly severed by nginx. This also leads to all files in the repo
being served by nginx though. This is not really desirable. The example
files, design files, mockups etc. should not be reachable via the
website, but they should be in the repo.

Therefore, all source files that are supposed to be served are moved
into the `src` directory. The nginx config can now be very simple
changed to only serve files from the src directory.